projects/agentic-ai/smart-claims-processor/requirements.txt
# Core LangGraph + LangChain
langgraph>=1.1.6
langgraph-checkpoint-sqlite>=3.0.3      # durable HITL interrupt/resume across restarts
langchain>=1.2.15
langchain-core>=1.2.28
langchain-google-genai>=4.2.1           # uses google-genai

# CrewAI (fraud detection crew - @tool decorator is in core crewai)
crewai>=1.14.1
# litellm is pulled transitively by crewai with compatible python-dotenv pins

# LLM Providers (pluggable - see src/llm.py + configs/base.yaml)
google-genai>=1.73.0                    # Google Gemini SDK
langchain-groq>=1.1.2

# Data validation
pydantic>=2.11.9,<2.12.0                 # pinned: crewai 1.14.x requires <2.12

# Web API (FastAPI backend + React frontend)
fastapi>=0.135.3
uvicorn[standard]>=0.44.0
sqlmodel>=0.0.38
python-jose[cryptography]>=3.5.0
bcrypt>=4.2.0,<5.0.0                    # bcrypt 5.x is a major bump; pin to 4.x for passlib compat
email-validator>=2.3.0

# Agent Memory (ChromaDB + HuggingFace embeddings)
chromadb>=1.1.0,<1.2.0                   # pinned: crewai 1.14.x requires <1.2
sentence-transformers>=4.0.0,<6.0.0     # 5.x OK but cap to avoid future breaks with torch/onnxruntime

# Configuration
pyyaml>=6.0.3
python-dotenv>=1.1.1,<1.2.0              # pinned: crewai 1.14.x requires <1.2

# Testing
pytest>=8.3.0,<10.0.0
